Always use a random server port when running `EngineContext`
